requirements.txt
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
defusedxml==0.7.1
dnslib==0.9.26
dnspython==2.7.0
Flask==3.1.0
flask-cors==5.0.1
google-api-core==2.24.2
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.1
googlesearch-python==1.3.0
gunicorn==22.0.0
h11==0.14.0
html5lib==1.1
httplib2==0.22.0
humanfriendly==10.0
idna==3.10
ipwhois==1.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mechanize==0.4.10
oauthlib==3.2.2
outcome==1.3.0.post0
phonenumbers==8.13.55
proto-plus==1.26.1
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyfiglet==1.0.2
pyparsing==3.2.1
PySocks==1.7.1
pyTelegramBotAPI==4.26.0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scapy==2.6.1
selenium==4.30.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
telebot==0.0.5
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wget==3.2
wikipedia==1.4.0
wsproto==1.2.0
pillow==10.3.0
openai==0.28
flask-limiter
duckduckgo-search==2.9.5
scikit-learn